modifing

+++ day07/ex01/iter.hpp
@@ -6,7 +6,7 @@
 /*   By: obelkhad <obelkhad@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/09 10:58:04 by obelkhad          #+#    #+#             */
-/*   Updated: 2022/10/09 16:18:07 by obelkhad         ###   ########.fr       */
+/*   Updated: 2022/10/12 19:55:42 by obelkhad         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,9 +16,9 @@
 
 #include <iostream>
 
-template <class T> void print(T arr[], int length)
+template <class T> void print(T arr[], unsigned int length)
 {
-	for (int i = 0; i < length; i++)
+	for (size_t i = 0; i < length; i++)
 		std::cout << "[]: " << arr[i] << std::endl;
 }
 template <class T> void inc(T &elem)
@@ -34,9 +34,9 @@ template <class T> void _32(T &elem)
 {
 	elem -= 32;
 }
-template <class T> void iter(T arr[], int length, void xFunc(T &))
+template <class T> void iter(T arr[], unsigned int length, void xFunc(T &))
 {
-	for (int i = 0; i < length; i++)
+	for (size_t i = 0; i < length; i++)
 		xFunc(arr[i]);
 }
 // template <class T, class F> void iter(T arr[], int length, F xFunc)
